Fix msvc windows builds

+++ vk_video_decoder/include/vkvideo_parser/PictureBufferBase.h
@@ -64,7 +64,7 @@ public:
     vkPicBuffBase()
         : m_refCount(0)
         , m_picIdx(-1)
-        , m_displayOrder(-1)
+        , m_displayOrder((uint32_t)-1)
         , m_decodeOrder(0)
         , m_timestamp(0)
         , m_presentTime(0)
